refactor(data-explorer): extract coordinators from panel
Extract three coordinators from DataExplorerPanelComposed to separate
concerns and reduce file size (2182 → 1016 lines):

- DataExplorerMetadataLoader: entity/attribute loading and caching
- DataExplorerExportImportCoordinator: export/import file operations
- VisualQueryBuilderCoordinator: query state (columns, filters, sort)

Panel remains orchestrator, delegating via typed context interfaces.
